updated code to 0.232

+++ mcs/src/src/emu/bookkeeping.cs
@@ -70,13 +70,13 @@ namespace mame
         -------------------------------------------------*/
         public void coin_counter_w(int num, int on)
         {
-            if (num >= std.size(m_coin_count))
+            if (num >= (int)std.size(m_coin_count))
                 return;
 
             /* Count it only if the data has changed from 0 to non-zero */
             if (on != 0 && (m_lastcoin[num] == 0))
                 m_coin_count[num]++;
-            m_lastcoin[num] = (UInt32)on;
+            m_lastcoin[num] = (u32)on;
         }
 
 
@@ -90,9 +90,9 @@ namespace mame
         -------------------------------------------------*/
         void coin_lockout_w(int num, int on)
         {
-            if (num >= std.size(m_coinlockedout))
+            if (num >= (int)std.size(m_coinlockedout))
                 return;
-            m_coinlockedout[num] = (UInt32)on;
+            m_coinlockedout[num] = (u32)on;
         }
 
         // return current lockout state for a particular coin
@@ -102,7 +102,7 @@ namespace mame
         -------------------------------------------------*/
         public int coin_lockout_get_state(int num)
         {
-            if (num >= std.size(m_coinlockedout))
+            if (num >= (int)std.size(m_coinlockedout))
                 return 0;
             return (int)m_coinlockedout[num];
         }
@@ -115,7 +115,7 @@ namespace mame
         -------------------------------------------------*/
         public void coin_lockout_global_w(int on)
         {
-            for (int i = 0; i < std.size(m_coinlockedout); i++)
+            for (int i = 0; i < (int)std.size(m_coinlockedout); i++)
                 coin_lockout_w(i, on);
         }
